@@ -9,32 +9,24 @@ repositories {
     maven { url 'https://oss.sonatype.org/content/repositories/ksoap2-android-releases' }
 }
 
-// Create a variable called keystorePropertiesFile, and initialize it to your
-// keystore.properties file, in the rootProject folder.
 def keystorePropertiesFile = rootProject.file("./key/keystore.properties")
-
-// Initialize a new Properties() object called keystoreProperties.
 def keystoreProperties = new Properties()
-
-// Load your keystore.properties file into the keystoreProperties object.
 keystoreProperties.load(new FileInputStream(keystorePropertiesFile))
 
-
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
     flavorDimensions "default"
 
     defaultConfig {
         applicationId "net.atebtech.dublin"
         minSdkVersion 18
-        targetSdkVersion 27
-        versionCode 1
-        versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        targetSdkVersion 28
+        versionCode 2
+        versionName "1.1"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         buildConfigField "boolean", "MOCK", "false"
         buildConfigField "boolean", "MOCK_MAP", "true"
-
-//        vectorDrawables.useSupportLibrary = true
+        multiDexEnabled true
     }
 
     signingConfigs {
