@@ -1,7 +1,7 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.2.51'
+    ext.kotlin_version = '1.3.10'
     repositories {
         google()
         jcenter()
@@ -10,10 +10,10 @@ buildscript {
         }
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.1.3'
+        classpath 'com.android.tools.build:gradle:3.2.1'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        classpath 'com.google.gms:google-services:4.0.1'
-        classpath 'io.fabric.tools:gradle:1.25.4'
+        classpath 'com.google.gms:google-services:4.2.0'
+        classpath 'io.fabric.tools:gradle:1.26.1'
         classpath 'com.google.firebase:firebase-plugins:1.1.5'
     }
 }
@@ -22,6 +22,10 @@ allprojects {
     repositories {
         google()
         jcenter()
+
+        maven {
+            url 'https://maven.google.com/'
+        }
     }
 }
 
