@@ -1,11 +1,12 @@
 package net.dublin.bus.ui.view.main
 
+import android.annotation.SuppressLint
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import android.support.design.widget.BottomNavigationView
-import android.support.v4.app.Fragment
-import android.support.v7.app.AppCompatActivity
+import com.google.android.material.bottomnavigation.BottomNavigationView
+import androidx.fragment.app.Fragment
+import androidx.appcompat.app.AppCompatActivity
 import android.util.Log
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
@@ -65,7 +66,7 @@ class MainActivity : AppCompatActivity() {
         initExtra()
 
         main_navigation.setOnNavigationItemSelectedListener(mOnNavigationItemSelectedListener)
-        BottomNavigationViewHelper.disableShiftMode(main_navigation)
+        //BottomNavigationViewHelper.disableShiftMode(main_navigation)
 
         main_search_view.setOnClickListener {
             SearchActivity.navigate(this)
@@ -82,6 +83,7 @@ class MainActivity : AppCompatActivity() {
         checkFavourite = intent.getBooleanExtra(EXTRA_CHECK_FAVOURITE, false)
     }
 
+    @SuppressLint("CheckResult")
     private fun checkFavourite() {
         if (checkFavourite) {
             val repository = StopRepository(application)
