@@ -1,6 +1,6 @@
 package net.dublin.bus.ui.view.realtime
 
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
