@@ -1,11 +1,11 @@
 package net.dublin.bus.ui.view.route.detail.map
 
 import android.annotation.SuppressLint
-import android.arch.lifecycle.Observer
-import android.arch.lifecycle.ViewModelProviders
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProviders
 import android.location.Location
 import android.os.Bundle
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -97,7 +97,7 @@ class RouteDetailMapFragment : Fragment(), OnMapReadyCallback, LocationRequestWr
     }
 
     private fun setUpMapIfNeeded() {
-        if (mSupportMapFragment == null) {
+        if (mSupportMapFragment === null) {
             val fm = childFragmentManager
             mSupportMapFragment = fm.findFragmentById(R.id.route_detail_map_view) as SupportMapFragment
             mSupportMapFragment?.getMapAsync(this)
