@@ -1,7 +1,8 @@
 package net.dublin.bus.ui.view.splash
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
+import com.google.firebase.FirebaseApp
 import net.dublin.bus.ui.view.main.MainActivity
 
 class SplashActivity : AppCompatActivity() {
