@@ -6,8 +6,8 @@ import android.content.Intent
 import android.graphics.Bitmap
 import android.os.Build
 import android.os.Bundle
-import android.support.annotation.NonNull
-import android.support.v7.app.AppCompatActivity
+import androidx.annotation.NonNull
+import androidx.appcompat.app.AppCompatActivity
 import android.text.TextUtils
 import android.view.View
 import android.webkit.WebResourceError
