@@ -1,9 +1,11 @@
 package net.dublin.bus
 
-import android.app.Application
+import androidx.multidex.MultiDexApplication
+import com.google.firebase.FirebaseApp
 
-class App : Application() {
+class App : MultiDexApplication() {
     override fun onCreate() {
         super.onCreate()
+        //FirebaseApp.initializeApp(this)
     }
 }
