@@ -1,12 +1,12 @@
 package net.dublin.bus.data.route.db
 
-import android.arch.lifecycle.LiveData
-import android.arch.persistence.room.Dao
-import android.arch.persistence.room.Insert
-import android.arch.persistence.room.OnConflictStrategy
-import android.arch.persistence.room.Query
+import androidx.lifecycle.LiveData
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
 import net.dublin.bus.model.Route
-import android.arch.persistence.room.Transaction
+import androidx.room.Transaction
 
 @Dao
 interface RouteDao {
