@@ -1,7 +1,7 @@
 package net.dublin.bus.data.stop.db
 
-import android.arch.lifecycle.LiveData
-import android.arch.persistence.room.*
+import androidx.lifecycle.LiveData
+import androidx.room.*
 import net.dublin.bus.model.Stop
 
 @Dao
