@@ -1,7 +1,7 @@
 package net.dublin.bus.model
 
-import android.arch.persistence.room.Entity
-import android.arch.persistence.room.Ignore
+import androidx.room.Entity
+import androidx.room.Ignore
 
 @Entity(tableName = "recent", primaryKeys = ["number", "type"])
 data class Recent(
