@@ -1,7 +1,7 @@
 package net.dublin.bus.ui.view.route.detail
 
-import android.arch.lifecycle.ViewModel
-import android.arch.lifecycle.ViewModelProvider
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 import net.dublin.bus.data.route.repository.RouteRepository
 
 class RouteDetailViewModelFactory(private val repository: RouteRepository) : ViewModelProvider.NewInstanceFactory() {
