@@ -1,12 +1,11 @@
 package net.dublin.bus.ui.view.route.detail.list
 
-import android.arch.lifecycle.Observer
-import android.arch.lifecycle.ViewModelProviders
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProviders
 import net.dublin.bus.data.route.repository.RouteRepository
 import net.dublin.bus.model.Stop
 import net.dublin.bus.ui.view.route.detail.RouteDetailViewModel
 import net.dublin.bus.ui.view.route.detail.RouteDetailViewModelFactory
-
 
 class RouteDetailPresenter(private val view: RouteDetailContract.View) : RouteDetailContract.Presenter {
 
